seoul_energy/requirements.txt
numpy
pandas
fastapi
uvicorn
supabase
sqlalchemy
google-genai
scikit-learn
python-dotenv
psycopg2-binary
pydantic-settings
pymupdf
pillow